requirements.txt
Django
djangorestframework
psycopg2-binary
celery
redis
django-redis
Pillow
pytest
pytest-django
python-dotenv